2019-2-22 echarts图表待

+++ src/main/java/com/tansuo365/test1/bean/user/User.java
@@ -12,13 +12,16 @@ import java.io.Serializable;
 
 /*管理员*/
 
+/**
+ * TODO 待加入生成时间,update时间
+ */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class User implements Serializable,MyLoginInstance {
 
-	private Long id;
+	private Integer id;
 
 	private String name;
